+++ protocols/request-response/src/lib.rs
@@ -28,7 +28,7 @@
 //! over the actual messages being sent, which are defined in terms of a
 //! [`Codec`]. Creating a request/response protocol thus amounts
 //! to providing an implementation of this trait which can then be
-//! given to [`Behaviour::new`]. Further configuration options are
+//! given to [`Behaviour::with_codec`]. Further configuration options are
 //! available via the [`Config`].
 //!
 //! Requests are sent using [`Behaviour::send_request`] and the
@@ -38,6 +38,14 @@
 //! Responses are sent using [`Behaviour::send_response`] upon
 //! receiving a [`Message::Request`] via
 //! [`Event::Message`].
+//!
+//! ## Predefined codecs
+//!
+//! In case your message types implement [`serde::Serialize`] and [`serde::Deserialize`],
+//! you can use two predefined behaviours:
+//!
+//! - [`cbor::Behaviour`] for CBOR-encoded messages
+//! - [`json::Behaviour`] for JSON-encoded messages
 //!
 //! ## Protocol Families
 //!
@@ -58,8 +66,12 @@
 
 #![cfg_attr(docsrs, feature(doc_cfg, doc_auto_cfg))]
 
+#[cfg(feature = "cbor")]
+pub mod cbor;
 mod codec;
 mod handler;
+#[cfg(feature = "json")]
+pub mod json;
 
 pub use codec::Codec;
 pub use handler::ProtocolSupport;
@@ -330,11 +342,24 @@ where
 
 impl<TCodec> Behaviour<TCodec>
 where
+    TCodec: Codec + Default + Clone + Send + 'static,
+{
+    /// Creates a new `Behaviour` for the given protocols and configuration, using [`Default`] to construct the codec.
+    pub fn new<I>(protocols: I, cfg: Config) -> Self
+    where
+        I: IntoIterator<Item = (TCodec::Protocol, ProtocolSupport)>,
+    {
+        Self::with_codec(TCodec::default(), protocols, cfg)
+    }
+}
+
+impl<TCodec> Behaviour<TCodec>
+where
     TCodec: Codec + Clone + Send + 'static,
 {
     /// Creates a new `Behaviour` for the given
     /// protocols, codec and configuration.
-    pub fn new<I>(codec: TCodec, protocols: I, cfg: Config) -> Self
+    pub fn with_codec<I>(codec: TCodec, protocols: I, cfg: Config) -> Self
     where
         I: IntoIterator<Item = (TCodec::Protocol, ProtocolSupport)>,
     {
@@ -682,7 +707,7 @@ where
     TCodec: Codec + Send + Clone + 'static,
 {
     type ConnectionHandler = Handler<TCodec>;
-    type OutEvent = Event<TCodec::Request, TCodec::Response>;
+    type ToSwarm = Event<TCodec::Request, TCodec::Response>;
 
     fn handle_established_inbound_connection(
         &mut self,
@@ -755,8 +780,9 @@ where
             FromSwarm::ExpiredListenAddr(_) => {}
             FromSwarm::ListenerError(_) => {}
             FromSwarm::ListenerClosed(_) => {}
-            FromSwarm::NewExternalAddr(_) => {}
-            FromSwarm::ExpiredExternalAddr(_) => {}
+            FromSwarm::NewExternalAddrCandidate(_) => {}
+            FromSwarm::ExternalAddrExpired(_) => {}
+            FromSwarm::ExternalAddrConfirmed(_) => {}
         }
     }
 
@@ -877,7 +903,7 @@ where
         &mut self,
         _: &mut Context<'_>,
         _: &mut impl PollParameters,
-    ) -> Poll<ToSwarm<Self::OutEvent, THandlerInEvent<Self>>> {
+    ) -> Poll<ToSwarm<Self::ToSwarm, THandlerInEvent<Self>>> {
         if let Some(ev) = self.pending_events.pop_front() {
             return Poll::Ready(ev);
         } else if self.pending_events.capacity() > EMPTY_QUEUE_SHRINK_THRESHOLD {